fix(select): Fix bug in the way the UCB1 was calculated on state brought in from the Transposition t
When calculating the UCB1 for a node, the current algorithm uses the parent node's visit count as a
proxy for the total of the siblings' visit count. This creates a problem when an existing state is
brought in from the transposition table. At that moment, the parent's visit count is no longer
reprensentatiove of the sum of its children visit count. Fix the algorithm so that it uses the
correct number.