feat: Remove enhancements to streamline the API
Most enhancements to the algorithm that were implemented actually reduced the quality of the output
in most of our tests. It looks like the  benefits of the enhancements do not outweight their
additional processing time cost. Remove all enhancements save for the transposition table. This will
make the API cleaner and configuration less of a hassle.